+++ source/toolkits/SignalTk.h
@@ -3,11 +3,12 @@
 
 #include <string>
 
+class ProgArgs; // forward declaration
 
 class SignalTk
 {
 	public:
-		static void registerFaultSignalHandlers();
+		static void registerFaultSignalHandlers(const ProgArgs& progArgs);
 		static bool blockInterruptSignals();
 		static bool unblockInterruptSignals();
 		static std::string logBacktrace();